PRE_Bigru/requirements.txt
tensorflow==1.14.0
numpy
bert_serving
pyyaml #yaml
pymongo
tqdm